requirements.txt
ase>=3.22.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
fairchem-core>=1.0.0
torch>=1.11.0
scipy>=1.7.0
pandas>=1.3.0
pyyaml>=6.0
tqdm>=4.62.0
